Cleanup.

+++ htp/htp.h
@@ -398,12 +398,7 @@ struct htp_cfg_t {
     htp_hook_t *hook_request_line;
 
     /** Request headers hook, invoked after all request headers are seen. */
-    htp_hook_t *hook_request_headers;
-
-    /** Raw request headers hook. This hook may be invoked twice if there
-     *  are any trailer headers in a request.
-     */
-    //htp_hook_t *hook_request_headers_raw;
+    htp_hook_t *hook_request_headers;   
 
     /** Request body data hook, invoked every time body data is available. Chunked data
      *  will be dechunked and compressed data will be decompressed (not implemented at present)
@@ -851,7 +846,7 @@ struct htp_tx_t {
     /** How many request header lines have been included in the raw
      *  buffer (above).
      */
-    int request_headers_raw_lines;
+    size_t request_headers_raw_lines;
 
     /** Request transfer coding: IDENTITY or CHUNKED. Only available on requests that have bodies. */
     int request_transfer_coding;
@@ -985,7 +980,6 @@ htp_cfg_t *htp_config_create();
 void htp_config_register_transaction_start(htp_cfg_t *cfg, int (*callback_fn)(htp_connp_t *));
 void htp_config_register_request_line(htp_cfg_t *cfg, int (*callback_fn)(htp_connp_t *));
 void htp_config_register_request_headers(htp_cfg_t *cfg, int (*callback_fn)(htp_connp_t *));
-//void htp_config_register_request_headers_raw(htp_cfg_t *cfg, int (*callback_fn)(htp_tx_data_t *));
 void htp_config_register_request_body_data(htp_cfg_t *cfg, int (*callback_fn)(htp_tx_data_t *));
 void htp_config_register_request_trailer(htp_cfg_t *cfg, int (*callback_fn)(htp_connp_t *));
 void htp_config_register_request(htp_cfg_t *cfg, int (*callback_fn)(htp_connp_t *));
@@ -1147,8 +1141,6 @@ char *htp_tx_progress_as_string(htp_tx_t *tx);
 bstr *htp_unparse_uri_noencode(htp_uri_t *uri);
 
 int htp_resembles_response_line(htp_tx_t *tx);
-
-// int htp_req_run_hook_request_headers_raw(htp_connp_t *connp, size_t first, size_t last);
 
 bstr *htp_tx_get_request_headers_raw(htp_tx_t *tx);
